Arduino Scripts
Three Arduino programs. 
'frequencysweep.ino' runs the pendulum through a sequence of increasing forcing frequencies and records angular response, and sends this information to dataextractor.py for collection.
'sineendoverend.ino' runs the pendulum at the correct frequency for a steady state in which the pendulum spins about the pivot.
'zerofinder.ino' uses the built-in limit switches of the track to determine the middle, and runs this routine multiple times to characterise how consistent the motor step number is.